app/requirements.txt
pip
Django<2.1
psycopg2
python-memcached
django-debug-toolbar
celery
django-celery-beat
django-celery-results
djangorestframework
djangorestframework-filters
django-filter
django-guardian
channels_api
channels
django-rest-swagger
djangorestframework-hstore
markdown
pexpect
Pillow
PyYAML
requests
django-admin-json-editor
django-cors-headers
django-cacheback[celery]
youtube-dl
gunicorn
git+https://github.com/django-recurrence/django-recurrence.git#egg=django_recurrence